Get scheduling working again
I reimplemented an account model that stores the identifier and
app_password (encrypted). Storing app_password is not ideal, but
it's not something we want to pass into the job params because that
logs out as plain text.

At least with rails encryption we get app_password log filtering out
of the box. And if an attacker gets into the DB, they're unable to
read the app_password.